feat: firebase reset
