+++ client/src/lib/firebase.ts
@@ -1,8 +1,16 @@
 'use client'
+import {
+  type Analytics,
+  type EventNameString,
+  type EventParams,
+  logEvent,
+} from 'firebase/analytics'
 // Import the functions you need from the SDKs you need
 import { initializeApp } from 'firebase/app'
-import { getAuth, type Auth } from 'firebase/auth'
+import { getAuth } from 'firebase/auth'
 import { type Firestore, getFirestore } from 'firebase/firestore'
+import { getStorage } from 'firebase/storage'
+
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
@@ -17,9 +25,19 @@ const firebaseConfig = {
   appId: '1:199379225695:web:50c2d3d4149239c7e71fea',
   measurementId: 'G-680CGPQBTE',
 }
-
-// Initialize Firebase
 const app = initializeApp(firebaseConfig)
-export const auth: Auth = getAuth(app)
-export const db: Firestore = getFirestore(app)
-export default app
+const auth = getAuth(app)
+const db: Firestore = getFirestore(app)
+const analytics: Analytics | null = null
+const storage = getStorage(app)
+
+export const fireAnalytics = (
+  eventName: EventNameString,
+  eventParams?: EventParams,
+) => {
+  if (analytics) {
+    logEvent(analytics, eventName as string, eventParams)
+  }
+}
+
+export { auth, db, analytics, storage }
